+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Revision history for finale
 
+## 0.1.1.0 -- 2022-05-27
+
+New Status: Starting. You can now position the ball before start playing
+Collision with base have been improved
+Collision with base corner implemented
+Base length is now dependant on width
+Bricks length is now dependant on width
+Smart walls height is now dependant on height
+
 ## 0.1.0.0 -- 2022-05-22
 
 * First version. Released on an unsuspecting world.